+++ Kernel/System/TemplateGenerator.pm
@@ -1676,13 +1676,15 @@ sub _Replace {
     $HashGlobalReplace->( $Tag, %Ticket, %DynamicFieldDisplayValues );
 
     # OTRS_TICKET_LAST_ARTICLE_ID
-    my $ArticleObject     = $Kernel::OM->Get('Kernel::System::Ticket::Article');
-    my @LastTicketArticle = $ArticleObject->ArticleList(
-        TicketID => $Ticket{TicketID},
-        OnlyLast => 1,
-    );
-    my $LastTicketArticleID = @LastTicketArticle ? $LastTicketArticle[0]->{ArticleID} : '';
-    $Param{Text} =~ s/$Start OTRS_TICKET_LAST_ARTICLE_ID $End/$LastTicketArticleID/gixms;
+    my $ArticleObject = $Kernel::OM->Get('Kernel::System::Ticket::Article');
+    if ( $Ticket{TicketID} ) {
+        my @LastTicketArticle = $ArticleObject->ArticleList(
+            TicketID => $Ticket{TicketID},
+            OnlyLast => 1,
+        );
+        my $LastTicketArticleID = @LastTicketArticle ? $LastTicketArticle[0]->{ArticleID} : '';
+        $Param{Text} =~ s{$Start OTRS_TICKET_LAST_ARTICLE_ID $End}{$LastTicketArticleID}gixms;
+    }
 
     # COMPAT
     $Param{Text} =~ s/$Start OTRS_TICKET_ID $End/$Ticket{TicketID}/gixms;